Some bug fixes, and now magic missile resistance is MR_MAGICMISSILE instead of hard-coded.

+++ include/monst.h
@@ -78,7 +78,7 @@ struct monst {
     uchar m_ap_type;      /* what mappearance is describing, m_ap_types */
 
     schar mtame;                /* level of tameness, implies peaceful */
-    unsigned short mextrinsics; /* low 8 correspond to mresists */
+    unsigned long mextrinsics; /* low 8 correspond to mresists */
     int mspec_used;             /* monster's special ability attack timeout */
 
     Bitfield(female, 1);      /* is female */
